1. Fix the flashback of the clicked song in the "History List" 2. The hysteresis caused by fast music playback 3. Add a hint to copy in the song details

+++ app/src/main/java/top/geek_studio/chenlongcould/musicplayer/Values.java
@@ -20,8 +20,8 @@ public interface Values {
 	/**
 	 * permission RequestCode
 	 */
-	int REQUEST_WRITE_EXTERNAL_STORAGE = 60;
-	int MAX_HEIGHT_AND_WIDTH = 100;
+	byte REQUEST_WRITE_EXTERNAL_STORAGE = 60;
+	byte MAX_HEIGHT_AND_WIDTH = 100;
 
 	/**
 	 * sharedPrefs tag
